requirements.txt
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
coloredlogs==15.0.1
Cython==3.2.4
dtw-python==1.7.4
exceptiongroup==1.3.1
ffmpeg-python==0.2.0
filelock==3.19.1
flatbuffers==25.12.19
fsspec==2025.10.0
future==1.0.0
h11==0.16.0
hanlp==2.1.3
hanlp_common==0.0.23
hanlp_downloader==0.0.25
hanlp_trie==0.0.5
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==0.36.2
humanfriendly==10.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
llvmlite==0.43.0
MarkupSafe==3.0.3
more-itertools==10.8.0
mpmath==1.3.0
networkx==3.2.1
numba==0.60.0
numpy==2.0.2
nvidia-ml-py==13.590.48
ollama==0.6.1
onnxruntime==1.19.2
openai-whisper==20250625
packaging==26.0
phrasetree==0.0.9
protobuf==6.33.5
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pynvml==13.0.1
pysrt==1.1.2
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
safetensors==0.7.0
scipy==1.13.1
sentencepiece==0.2.1
silero-vad==6.2.0
six==1.17.0
soundfile==0.13.1
sympy==1.14.0
termcolor==3.1.0
tiktoken==0.12.0
tokenizers==0.22.2
toposort==1.5
torch==2.8.0
torchaudio==2.8.0
tqdm==4.67.3
transformers==4.57.6
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
whisper-timestamped==1.15.9
